backend/requirements.txt
# Python 3.11–3.14 supported (3.14 needs pydantic >= 2.10 for pre-built wheels).
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.26.0
sendgrid==6.11.0
pydantic[email]>=2.10,<3
pydantic-settings>=2.6
python-multipart==0.0.9
